Almost done with account creator class
I need to make some optimization on methods that deal with ID numbers and card numbers, The user input should be treated as a string and get saved as a string otherwise they will occupy a huge amount of memory.

+++ ATMSimulator/Account.cs
@@ -8,13 +8,13 @@ namespace ATMSimulator
 {
     internal class Account
     {
-        public int IDnum { get; set; }
+        public ulong IDnum { get; set; }
         public  string Name { get; set; }
         public string LastName { get; set; }
         public int CardNum { get; set; }
         public int Password { get; set; }
 
-        public Account(int iDnum, string name, string lastName, int cardNum, int password)
+        public Account(ulong iDnum, string name, string lastName, int cardNum, int password)
         {
             IDnum = iDnum;
             Name = name;
